Migrate from Angular 4 to Angular 5
This gets us up to Angular 5.2.5 and everything appears to work.

I had to hand upgrade the version of `"@angular/flex-layout"`
in `package.json` to `"^2.0.0-beta.10"`. Version `beta.9` (which is
what we had before) didn't work with KK's new E2E tests, but upping
to `beta.10` appears to fix that.