Add delay so E2E tests pass
This comments out the delay at the top of these E2E tests, which then
causes one of the other tests to fail because we jump to the expect
before the user has been fully added and updated. I had to add a 10
second delay to get this to pass consistently.