Fix use of mat-list
This commit updates the code in user.component.html to use an Angular Material Design mat-list. I'm not convinced this is the best use of mat-list, but the documentation on mat-list says "<mat-list> is a container component that wraps and formats a series of line items. As the base list component, it provides Material Design styling, but no behavior of its own." I think my previous commit was relying on it to provide behavior it wasn't providing, but I don't feel completely confident. This might not matter because it really does not seem to show up anywhere in the project, but the tests are passing now. Closes issue #11. Maybe.