Fix use of mat-list again, but differently
This commit updates the code in user.component.html to use an Angular Material Design mat-list and updates the karma tests that check that the world still works. I think this is a more appropriate way to use mat-list than my most recent commit before this. The problem I had that the previous commit didn't recognize or fix was that the karma tests were not working with mat-list. Importing CustomModule fixes that.