Update angular (CLI) (#9)
* Upgrade Angular (CLI)

The goal here was to upgrade Angular CLI to avoid a bunch of security
warnings. I ended up having to upgrade quite a lot to get things to
work, and I think we've effectively upgraded to Angular 5 at this point.

* Remove all references to `MATERIAL_COMPATIBILITY_MODE`

This got removed starting with Angular 5, and we didn't need it anymore
since we had changed all our `Md-` components to `Mat-` components.

* Update Gradle tasks

This fixes some problems caused by changes to the flags to `ng`, sorts
out some warnings, and updates the project name.

* Reduce compatibility warnings

This eliminates a few library compatibility
warnings caused by `flex-layout`. We still have
two which it looks like we're stuck with for now.
I don't think that `@angular/flex-layout` is
getting a lot of development attention, so we
may have to live with some of these problems if
we're going to continue to use it.

* Roll node back to v8

I can't get Travis to deal with versions of
node higher than 8, so I had to roll that back 
to get our builds to
succeed. I think we could use `nvm` on Travis to
specify a more recent version, but this 
gets us over the current hump.